# Revive #27659: Revert "Revert "enable the SLP Vectorizer optimization pass by default""