@nanosoldier 2018-07-30
[Your benchmark job](https://github.com/JuliaLang/julia/pull/28344) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/b60231929b1990e804e4eb538807e35a0596107e/b760c78_vs_d7d5f11/report.md). cc @ararslan