@jrevels 2018-07-30
I'll investigate the perf regressions and report back. It's totally possible they're real; if so, do we still want to merge this?

@andreasnoack I'm seeing a [LinearAlgebra-y error on windows](https://ci.appveyor.com/project/JuliaLang/julia/build/1.0.29418/job/c6hadpaavg1fg50l#L2882) in this PR, I think it's unrelated but just want to check with you.