@jrevels 2018-07-30
I am seeing some reproducible regressions here, but they're smaller on my machine than reported by nanosoldier (max ~10% in the `skipmissing` case). 

That still seems worth merging to me; I'll do so tomorrow morning unless there are any objections.