@Keno 2018-07-30
> I am seeing some reproducible regressions here, but they're smaller on my machine than reported by nanosoldier (max ~10% in the skipmissing case).

It's quite possible that this varies over architectural features. What architecture did you test with?