@jrevels 2018-07-31
Shall I merge this? Not sure that the (potential) regressions on nanosoldier are worth holding back here, though if others want to test more they can