@Keno 2018-07-31
I'm ok with this, but rc1 just went, so this'll be in rc2. Hold for a few hours since we'll be branching release-0.7 shortly.