@jrevels 2018-07-31
> Hold for a few hours since we'll be branching release-0.7 shortly.

Okay, feel free to merge whenever. Otherwise I'll just merge tomorrow.